Extract Prompt Studio advanced layout controller
