+++ docs/development/issue-14-frontend-js-maintenance.md
@@ -208,6 +208,7 @@ web/js/
     canvas_forwarding.js
     serialization.js
     layout.js
+    advanced_layout_controller.js
     studio_textareas.js
     textarea.js
     wheel.js
@@ -241,19 +242,20 @@ Planned slices:
 9. `canvas_forwarding.js`
 10. `serialization.js`
 11. `layout.js`
-12. `studio_textareas.js`
-13. `textarea.js`
-14. `wheel.js`
-15. `dom.js`
-16. `extend_slots.js`
-17. `extend_slot_controls.js`
-18. `style.js`
-19. `tooltip.js`
-20. `widgets.js`
-21. `legend.js`
-22. `fields.js`
-23. `node_hooks.js`
-24. entry file slim-down
+12. `advanced_layout_controller.js`
+13. `studio_textareas.js`
+14. `textarea.js`
+15. `wheel.js`
+16. `dom.js`
+17. `extend_slots.js`
+18. `extend_slot_controls.js`
+19. `style.js`
+20. `tooltip.js`
+21. `widgets.js`
+22. `legend.js`
+23. `fields.js`
+24. `node_hooks.js`
+25. entry file slim-down
 
 Each PR should mostly move code and update imports. Behavior changes should be
 separate PRs unless the move reveals a confirmed bug that cannot be separated.
@@ -731,6 +733,7 @@ Use this in Issue #14 or follow-up PR tracking comments.
 [ ] canvas_forwarding.js extracted
 [ ] serialization.js extracted
 [ ] layout.js extracted
+[ ] advanced_layout_controller.js extracted
 [ ] studio_textareas.js extracted
 [ ] textarea.js extracted
 [ ] wheel.js extracted
